tweak: update reporting a little bit
