+++ src/lib/mod.rs
@@ -3,4 +3,3 @@ pub mod exception;
 pub mod flags;
 pub mod object;
 pub mod repl;
-pub mod telemetry;
